Refactor styles: reorganize CSS imports and move styles to dedicated files
