+++ responsive-recipe-website/src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route, Link } from 'react-router-dom';
 import RecipePage from './components/RecipePage.jsx';
 import CreateRecipe from './components/CreateRecipe.jsx';
 import RecipeDetails from './components/RecipeDetails.jsx';
-import './App.css';
+import './styles/App.css';
 
 function App() {
   return (
